DPAA2: FDT MDIO support
Rename memac_mdio_acpi.c to memac_mdio.c and add FDT support.  Extend
dpaa2_mc_fdt.c for the equivalent of the ACPI version to add supoprt
for getting the xref to mdio from dpamc.
Add a new mdio interface to get the phy location as with the multi-phy
MDIO bits we need to pass the phyloc information in to mii_attach.

This all needs a bit more unificationa and cleanup but works for the
copper ports so far.  We do not yet have any (generic) 'sff,sfp'
support in FreeBSD so that needs to come before enabling the sfp
bits.  Until then keeping sfpmode legacy in u-boot is probably best.

Hardware provided by:	Traverse Technologies (Ten64)